+++ cmd/bundle/deploy.go
@@ -3,6 +3,7 @@ package bundle
 import (
 	"github.com/databricks/bricks/bundle"
 	"github.com/databricks/bricks/bundle/phases"
+	"github.com/databricks/bricks/cmd/root"
 	"github.com/spf13/cobra"
 )
 
@@ -10,7 +11,7 @@ var deployCmd = &cobra.Command{
 	Use:   "deploy",
 	Short: "Deploy bundle",
 
-	PreRunE: ConfigureBundle,
+	PreRunE: root.MustConfigureBundle,
 	RunE: func(cmd *cobra.Command, args []string) error {
 		b := bundle.Get(cmd.Context())
 		return bundle.Apply(cmd.Context(), b, []bundle.Mutator{
