+++ cmd/bundle/validate.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/databricks/bricks/bundle"
 	"github.com/databricks/bricks/bundle/phases"
+	"github.com/databricks/bricks/cmd/root"
 	"github.com/spf13/cobra"
 )
 
@@ -12,7 +13,7 @@ var validateCmd = &cobra.Command{
 	Use:   "validate",
 	Short: "Validate configuration",
 
-	PreRunE: ConfigureBundle,
+	PreRunE: root.MustConfigureBundle,
 	RunE: func(cmd *cobra.Command, args []string) error {
 		b := bundle.Get(cmd.Context())
 		err := bundle.Apply(cmd.Context(), b, []bundle.Mutator{
